phytium_mci

Macro genmask_ull

macro_rules! genmask_ull {
    ($h:expr, $l:expr) => { ... };
}

Creates a contiguous bitmask for 64-bit values.

This macro generates a 64-bit mask with bits set from position l (low) to h (high), inclusive. Both positions are zero-indexed from the least significant bit.

Arguments

$h - High bit position (inclusive, 0-63)

$l - Low bit position (inclusive, 0-63)

Examples

assert_eq!(genmask_ull!(39, 21), 0x000000ffffe00000);
assert_eq!(genmask_ull!(63, 32), 0xFFFFFFFF00000000);

Notes

The high bit $h must be greater than or equal to the low bit $l

Both positions must be within 0-63 range for 64-bit masks