phytium_mci

Macro set_reg32_bits

macro_rules! set_reg32_bits {
    ($reg:expr, $a:expr, $b:expr) => { ... };
}

Prepares a value to be written to a bit field in a 32-bit register.

Takes a value and shifts it to the correct position for bits [a:b] of a register, where a is the high bit position and b is the low bit position.

Arguments

$reg - Value to write (will be shifted to position)

$a - High bit position (inclusive)

$b - Low bit position (inclusive)

Examples

let value = 0b1011u32; // Value to write
let prepared = set_reg32_bits!(value, 7, 4); // Prepare for bits 7:4
assert_eq!(prepared, 0b1011_0000u32); // Shifted to position