phytium_mci::mci

Struct MCI

pub struct MCI { /* private fields */ }

MCI (Memory Card Interface) hardware controller.

MCI represents the low-level hardware controller for SD/MMC operations on Phytium SoCs. It manages register access, DMA descriptor chains, PIO transfers, and interrupt handling.

Transfer Modes

DMA Mode: High-performance transfers using chained DMA descriptors

PIO Mode: Simple FIFO-based programmed I/O transfers

Timing

The controller supports multiple timing configurations for different clock speeds:

400 KHz (initialization)

25 MHz (default speed)

50 MHz (high speed)

208 MHz (SDR104 UHS-I)

IOPAD Integration

The controller optionally manages I/O pad configuration for signal delay tuning at high-speed clock rates.

Implementations

impl MCI

pub fn dma_int_set(&mut self)

Sets DMA interrupt enable bits

Enables receive, transmit, and fatal bus error interrupts for DMA transfers.

pub fn dump_dma_descriptor(&self, desc_in_use: u32)

Dumps DMA descriptor information for debugging

Arguments

desc_in_use - Number of descriptors currently in use

impl MCI

pub fn interrupt_mask_get(&self, tp: MCIIntrType) -> u32

Gets the SDIF controller interrupt mask

Arguments

tp - Interrupt type (General or DMA)

Returns

The current interrupt mask value

pub fn interrupt_mask_set(&self, tp: MCIIntrType, set_mask: u32, enable: bool)

Enables or disables SDIF controller interrupts

Arguments

tp - Interrupt type (General or DMA)

set_mask - Interrupt mask bits to modify

enable - True to enable interrupts, false to disable

impl MCI

pub const SWITCH_VOLTAGE: u32 = 11u32

Command index for voltage switching (CMD11)

impl MCI

MCI pub API

pub fn iopad_set(&mut self, iopad: IoPad)

Sets the I/O pad for signal delay tuning

Arguments

iopad - The I/O pad to use for delay configuration

pub fn iopad_take(&mut self) -> Option<IoPad>

Takes and returns the I/O pad, leaving None in its place

Returns

Some(IoPad) - If an I/O pad was set

None - If no I/O pad was configured

pub fn cur_cmd_set(&mut self, cmd: &MCICmdData)

Sets the current command being executed

Arguments

cmd - The command data to set as current

pub fn config_init(&mut self, config: &MCIConfig) -> Result<(), MCIError>

initialization SDIF controller instance

pub fn config_deinit(&mut self) -> Result<(), MCIError>

deinitialization SDIF controller instance

pub fn set_idma_list( &mut self, desc: &PoolBuffer, desc_num: u32, ) -> Result<(), MCIError>

Setup DMA descriptor for SDIF controller instance

pub fn clk_freq_set(&mut self, clk_hz: u32) -> Result<(), MCIError>

Set the Card clock freqency

pub fn dma_transfer( &mut self, cmd_data: &mut MCICmdData, ) -> Result<(), MCIError>

Start command and data transfer in DMA mode

pub fn poll_wait_dma_end( &mut self, cmd_data: &mut MCICmdData, ) -> Result<(), MCIError>

Wait DMA transfer finished by poll

pub fn pio_transfer(&self, cmd_data: &mut MCICmdData) -> Result<(), MCIError>

Start command and data transfer in PIO mode

pub fn poll_wait_pio_end( &mut self, cmd_data: &mut MCICmdData, ) -> Result<(), MCIError>

Wait PIO transfer finished by poll

pub fn restart(&self) -> Result<(), MCIError>

Reset controller from error state

pub fn register_dump(&self)

Dump all register value of SDIF instance

pub fn cmd_info_dump(cmd_data: &MCICmdData)

Dump command and data info